app/scraper/pokemon_scraper.py
"""
宝可梦数据爬取器 - 从神奇宝贝百科爬取数据
"""
import requests
from bs4 import BeautifulSoup
import time
from typing import List, Dict, Any


class PokemonScraper:
    """宝可梦数据爬取器"""
    
    BASE_URL = "https://wiki.52poke.com"
    
    def __init__(self, delay=1.0, max_retries=3, timeout=30):
        self.delay = delay
        self.max_retries = max_retries
        self.timeout = timeout
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        }
    
    def _fetch_page(self, url: str, retries: int = 0) -> str:
        """获取网页内容"""
        try:
            response = requests.get(url, headers=self.headers, timeout=self.timeout)
            response.raise_for_status_code()
            return response.text
        except Exception as e:
            if retries < self.max_retries:
                print(f"  重试获取 {url} (尝试 {retries + 1}/{self.max_retries})")
                time.sleep(self.delay * (retries + 1))
                return self._fetch_page(url, retries + 1)
            raise Exception(f"获取失败: {url} - {e}")
    
    def scrape_pokemon_list(self, max_count: int = None) -> List[Dict[str, Any]]:
        """爬取宝可梦列表"""
        print("开始爬取宝可梦列表...")
        
        url = f"{self.BASE_URL}/wiki/%E5%AE%9D%E5%8F%AF%E6%A2%A6%E5%88%97%E8%A1%A8%EF%BC%88%E6%8C%89%E5%85%A8%E5%9B%BD%E5%9B%BE%E9%89%B4%E7%BC%96%E5%8F%B7%E3%E7%BC%96%E5%8F%B7%EF%BC%88%E6%8C%89%E5%85%A8%E5%9B%BD%E5%9B%BE%E9%89%B4%E7%BC%96%E5%8F%B7%E3%E7%BC%96%E8%BF%9E0%E7%B5%AE%E9%98%A7%E8%AF%B1%E9%A1%98"
        
        html = self._fetch_page(url)
        soup = BeautifulSoup(html, 'html.parser')
        pokemon_list = []
        
        # 简化版本：从简单版页面解析
        # 这里需要根据实际网站结构进行解析
        # 示例数据
        return pokemon_list